app/ai_analyzer_service/requirements.txt
redis
rq
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-dotenv
google-generativeai
httpx
# google-cloud-aiplatform # If vertexai.preview.tokenization is used
